neo4j_runway/requirements.txt
graphviz==0.20.1
instructor==0.5.2
neo4j==5.8.0
openai==1.12.0
pandas==2.0.3
pydantic==2.6.1
pyyaml==6.0.1
streamlit==1.31.1

# pyingest
jmespath==0.10.0
numpy
python-dateutil==2.8.2
pytz==2023.4
# PyYAML==5.3.1
smart-open==4.0.1
urllib3==2.2.0